Allow a CNAME DNS-name in addition to the EFS ID
This change introduces additional logic which allows the user to specify
either a fully-qualified EFS DNS name, or a custom DNS name that
resolves to a fully-qualified EFS DNS name via a CNAME record.

The EFS DNS name will then be compared to the EFS name we would have
expected for the given EFS ID. If it doesn't match, we'll show the user
a readable error message for this to aid in troubleshooting.

Once the EFS DNS name has been verified, the rest of the logic is
untouched. This means that this change only impacts the parameter logic
at the very start and does not touch anything of the already tested,
more critical logic.

Important: the usage and internal logic to mount by EFS ID is unchanged,
making this change fully backwards compatible.

Fixes #9.