Mount by IP with TLS for clusters with DNS hostnames disabled
This introduces a 'dnshostnamesdisabled' option to supply during mount.
When supplied, the efs-utils package will lookup an IP address (using
AWS APIs) of a matching mount target and mount using the IP address.

An IP address is chosen by fetching the mount targets for the given
filesystem ID using AWS APIs, and filtering them by the subnet IDs of
the subnets of the network interfaces attached to the instance. The
first matching mount target IP address is chosen. If no such mount
target exists, return an error.

Using this option enables users to mount file systems in clusters with
DNS hostnames disabled, while still leveraging TLS.

Note that in order to use this feature:

- The host must have the botocore library installed
- The instance must have an IAM role with access to the
  'elasticfilesystem:DescribeMountTargets' API

https://docs.aws.amazon.com/efs/latest/ug/API_DescribeMountTargets.html