detect invalid private key
The `privateKey.pem` can become an empty file due to several reasons, but this case is falsely detected as an existing valid key.
Instead of just assuming an existing file means a valid key, at least there should be a check if the file is not empty.

There are several issues filed that cover this situation:

https://github.com/kubernetes-sigs/aws-efs-csi-driver/issues/683

https://github.com/kubernetes-sigs/aws-efs-csi-driver/issues/569
Signed-off-by: Samuel Lang <gh@lang-sam.de>