Disable journal entry fetch from systemctl call
systemctl has known performance issues with large numbers of journal files.

See https://github.com/systemd/systemd/issues/2460.

Changing to use `is-active` to improve call speed when systemctl is called. This
avoids first systemctl call issues that may result in high latency for the mount
as systemctl indexes all journal files. Given that mount.efs is only looking for
the exit code of the command and does not care about the journal entries, there
is no loss of usability by adding the flag.