Add warning if using old version
Perform a check in watchdog to see if newer version of
amazon-efs-utils is available on yum or github.  If so, log a warning to the
watchdog log file.  To disable this version check, we've added a
config value, enable_version_check.  If check fails,
we silently fail.

Also added Amazon Linux 2023 as supported distribution.