Check private key file size to skip generation
Nowadays, the private key generation function checks if the private
key file exists. However, if the openssl command that generates
the private key file is interrupted before writing the key, the
file is there but empty. Such behaviour has been confirmed locally.

That causes the following issue (https://github.com/kubernetes-sigs/aws-efs-csi-driver/issues/683)
when the EFS CSI driver pod is restarted. On the next start of the
pod in the node, it doesn't generate the private key because of
what is stated above.

With this change, if the file is empty it will be generated. No
need for the workaround with the initContainer to delete the file.