test: Change repo urls in eol debian9 build
Debian 9 is EOL and entered ELTS, which is not part of the Debian
project. Repository URLs changed from deb.archive.org /
security.debian.org to archive.debian.org, also 'stretch-updates' is not
available anymore and need to removed.

This commit changes the urls and removes 'updates' before installing
packages. It also ensures that 'git' is installed as dependency and the
temporary build folder added as safe git directory.

fixes: https://github.com/aws/efs-utils/issues/183