remove $ from all codeblocks
Having `$` at the beginning of lines in bash codeblocks makes copy and pasting a pain. We regularly spin up new ec2 instances and need to manually clone and build efs-utils from this repo. 

If the `$` were included to safeguard against users running pasted code without thinking, then there were many code blocks (also with sudo) where the lines did not begin with `$`. 

So I figured it was no a convention of the repository.