Regenerate Cargo.lock with rust 1.70.0
In the previous commit, Cargo.lock was updated with a more recent rust
toolchain, which generated a cargo lock with `version = 4`. This led to
build errors with customers using older rust versions. In this CR, we
Cargo.lock is regenerated using rust 1.70.0 so that a `version = 3`
Cargo.lock is generated.